+++ pymc3/sampling.py
@@ -30,6 +30,7 @@ import warnings
 import numpy as np
 import theano.gradient as tg
 from theano.tensor import Tensor
+import xarray
 
 from .backends.base import BaseTrace, MultiTrace
 from .backends.ndarray import NDArray
@@ -1520,9 +1521,9 @@ def sample_posterior_predictive(
 
     Parameters
     ----------
-    trace: backend, list, or MultiTrace
-        Trace generated from MCMC sampling. Or a list containing dicts from
-        find_MAP() or points
+    trace: backend, list, xarray.Dataset, or MultiTrace
+        Trace generated from MCMC sampling, or a list of dicts (eg. points or from find_MAP()),
+        or xarray.Dataset (eg. InferenceData.posterior or InferenceData.prior)
     samples: int
         Number of posterior predictive samples to generate. Defaults to one posterior predictive
         sample per posterior sample, that is, the number of draws times the number of chains. It
@@ -1556,6 +1557,23 @@ def sample_posterior_predictive(
         Dictionary with the variable names as keys, and values numpy arrays containing
         posterior predictive samples.
     """
+    if isinstance(trace, xarray.Dataset):
+        # grab posterior samples for each variable
+        _samples = {
+            vn : trace[vn].values
+            for vn in trace.keys()
+        }
+        # make dicts
+        points = []
+        for c in trace.chain:
+            for d in trace.draw:
+                points.append({
+                    vn : s[c, d]
+                    for vn, s in _samples.items()
+                })
+        # use the list of points
+        trace = points
+
     len_trace = len(trace)
     try:
         nchain = trace.nchains